A bit of cleanup...
There's no point doing the flexbox in IE, because its main purpose
is to let me bottom-align the bottom bumper without having to
specify an exact positioning in pixels. The reason I need to do
that, is because in mobile Safari and mobile Chrome, the hiding
and showing of the address bar can change the viewport size and
make bottom-aligned things disappear.

But a flexbox set to screen height remains the right size through
those changes, and a bottom-aligned bumper within the flexbox
stays bottom-aligned.