Trying to make the code more understandable
Moved the code for the top bumper to come first
(because that's where I intuitively seem to
expect it), and the bottom bumper on bottom.

Also adding some comments, and reversing
the &&-orders, so that I can better
conceptualize why this conditional
actually determines the right place
to make the bumpers start being fixed
position.