An even weirder Chrome workaround
In Chrome mobile on my Android phone, the kitten shows up as
transparent if I use a clip-path inset of 1px! If at least
one side of the inset is 2px or more, then the kitten is
fully hidden.

In case this is tied to some percentage of screensize, I'm
bumping it up to 10px, which hopefully will give some leeway.